requirements.txt
agno
youtube_transcript_api
streamlit
pillow
requests


        